mengubah dari psql ke mysql

+++ actix_web_server/src/models.rs
@@ -11,13 +11,13 @@ pub struct  InputDataExpenses {
 
 #[derive(Deserialize)]
 pub struct DeleteDataExpenses {
-    pub id: i32,
+    pub id: u64,
 }
 
 
 #[derive(Serialize, FromRow)]
 pub struct Expense {
-    pub id: i32,
+    pub id: u64,
     pub name: String,
     pub amount: i32,
     pub user_id: String,
@@ -26,7 +26,7 @@ pub struct Expense {
 
 #[derive(Serialize, FromRow)]
 pub struct LogData {
-    pub id: i32,
+    pub id: u64,
     pub location: String,
     pub created_at: Option<DateTime<Local>>,
 }
